% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarizar_venda_media.R
\name{sumarizar_venda_media}
\alias{sumarizar_venda_media}
\title{Sumarize a venda média pelo grupo especificado}
\usage{
sumarizar_venda_media(df, nome_coluna)
}
\arguments{
\item{df}{a base de vendas}

\item{nome_coluna}{coluna com grupos
na qual vai calcular a média de valor venda}
}
\value{
dataframe com as medias de vendas agrupadas por nome_coluna
}
\description{
Sumarize a venda média pelo grupo especificado
}
